# module4/webserver/swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "copy of the swagger PetStore",
        "title": "Swagger Rest API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/store/order": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "store"
                ],
                "summary": "Place an order for a pet",
                "operationId": "create-order-handler",
                "parameters": [
                    {
                        "description": "order placed for purchasing the pet",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/store.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/store.Order"
                        }
                    },
                    "400": {
                        "description": "Invalid Order"
                    }
                }
            }
        },
...
        "/store/order/{orderID}": {
            "get": {
                "description": "For valid response try integer IDs with value \u003e= 1 and \u003c= 10. Other values will generated exceptions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "store"
                ],
                "summary": "Place an order for a pet",
                "operationId": "get-order-handler",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of pet that needs to be fetched",
                        "name": "orderID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/store.Order"
                        }
                    },
                    "400": {
                        "description": "Invalid Order"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            },
            "delete": {
                "description": "For valid response try integer IDs with positive integer value. Negative or non-integer values will generate API errors",
                "tags": [
                    "store"
                ],
                "summary": "Place an order for a pet",
                "operationId": "delete-order-handler",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of pet that needs to be deleted",
                        "name": "orderID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            }
        },
        "/user": {
            "post": {
                "description": "This can only be done by the logged in user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Create user",
                "operationId": "user-create-handler",
                "parameters": [
                    {
                        "description": "Created user object",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.User"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Pet not found"
                    },
                    "default": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/createWithList": {
            "post": {
                "description": "This can only be done by the logged in user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Create with list user",
                "operationId": "user-create-with-list-handler",
                "parameters": [
                    {
                        "description": "List of user object",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/user.User"
                            }
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Pet not found"
                    },
                    "default": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/{Username}": {
            "get": {
                "description": "This can only be done by the logged in user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get user",
                "operationId": "user-get-by-username-handler",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The name that needs to be fetched",
                        "name": "Username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/user.User"
                        }
                    },
                    "400": {
                        "description": "Invalid username supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "put": {
                "description": "This can only be done by the logged in user.",
                "tags": [
                    "user"
                ],
                "summary": "Update user",
                "operationId": "user-update-by-username-handler",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The name that needs to be updated",
                        "name": "Username",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated user object",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.User"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid username supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "post": {
                "description": "This can only be done by the logged in user.",
                "tags": [
                    "user"
                ],
                "summary": "Delete user",
                "operationId": "user-delete-by-username-handler",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The name that needs to be deleted",
                        "name": "Username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid username supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "store.Order": {
            "type": "object",
            "properties": {
                "complete": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "petId": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "shipDate": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "user.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "userStatus": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}